✨ domain: relocate the real per-org provisioner into the operator, delete the stub
Move cert-manager.client, cloud-dns.client, org-domain.provisioner, its
factory and types from the control-plane (where they were dead — never
called) into apps/operator/src/cluster-tenants/internal, adapting to the
operator's CustomObjectsApi. The reconciler now builds and drives the REAL
DefaultOrgDomainProvisioner via _BuildOrgDomainProvisioner, replacing the
GatedOrgDomainProvisioner always-skip stub (deleted, including the
"intentionally unimplemented" comment).

Gating is now real capability detection: the Certificate manifest is built
and applied for real; an absent Certificate CRD short-circuits fail-closed,
the DNS side runs only when a Cloud DNS zone + ingress IP are configured,
and only when NEITHER side can act is the step reported skipped. Never throws.

Types split into org-domain-provisioner.types.ts (no interface+class in one
file). Operator config gains ingressIp/dnsManagedZone/certManager issuer +
namespace-prefix; @opencrane/contracts + @google-cloud/dns (optional) move to
the operator package; control-plane drops the dead copies, tests and barrel
re-exports.
