+++ apps/operator/package.json
@@ -13,10 +13,12 @@
   },
   "dependencies": {
     "@kubernetes/client-node": "^1.0.0",
+    "@opencrane/contracts": "workspace:*",
     "@opencrane/observability": "workspace:*",
     "pino": "^9.6.0"
   },
   "optionalDependencies": {
+    "@google-cloud/dns": "^4.0.0",
     "@google-cloud/storage": "^7.0.0"
   },
   "devDependencies": {
