+++ apps/operator/src/__tests__/fixtures.ts
@@ -17,6 +17,11 @@ export const defaultConfig: OpenClawTenantOperatorConfig = {
   tenantDefaultImage: "ghcr.io/opencrane/tenant:latest",
   defaultOpenclawVersion: "2026.6.9",
   ingressDomain: "opencrane.local",
+  ingressIp: "",
+  dnsManagedZone: "",
+  certManagerIssuerName: "opencrane-issuer",
+  certManagerIssuerKind: "ClusterIssuer",
+  clusterTenantNamespacePrefix: "opencrane-",
   ingressTlsEnabled: false,
   ingressTlsSecretName: "opencrane-wildcard-tls",
   gatewayPort: 18789,
