+++ apps/operator/src/__tests__/runtime-planes/drift-repairer.test.ts
@@ -15,6 +15,11 @@ function _buildConfig(overrides: Partial<OpenClawTenantOperatorConfig> = {}): Op
     tenantDefaultImage: "opencrane/tenant:latest",
     defaultOpenclawVersion: "2026.6.9",
     ingressDomain: "opencrane.local",
+    ingressIp: "",
+    dnsManagedZone: "",
+    certManagerIssuerName: "opencrane-issuer",
+    certManagerIssuerKind: "ClusterIssuer",
+    clusterTenantNamespacePrefix: "opencrane-",
     ingressTlsEnabled: false,
     ingressTlsSecretName: "opencrane-wildcard-tls",
     gatewayPort: 18789,
