+++ apps/operator/src/config.ts
@@ -32,6 +32,29 @@ export interface OpenClawTenantOperatorConfig
   /** Base domain for tenant ingress hostnames. */
   ingressDomain: string;
 
+  /**
+   * Cluster ingress external IP the per-org wildcard A records point at. Empty when
+   * unknown (e.g. on-prem or before the LoadBalancer IP is assigned); the per-org DNS
+   * side effect is then skipped and only the Certificate is applied.
+   */
+  ingressIp: string;
+
+  /**
+   * Cloud DNS managed-zone resource name (terraform `<zone>-zone`) the per-org A
+   * records are written into. Empty when the install is not on a Cloud DNS substrate;
+   * the per-org DNS side effect is then skipped.
+   */
+  dnsManagedZone: string;
+
+  /** cert-manager issuer name the per-org Certificate references. */
+  certManagerIssuerName: string;
+
+  /** Issuer kind for the per-org Certificate: `ClusterIssuer` (default) or `Issuer`. */
+  certManagerIssuerKind: "ClusterIssuer" | "Issuer";
+
+  /** Prefix applied to an org name to derive its bound namespace (`opencrane-<org>`). */
+  clusterTenantNamespacePrefix: string;
+
   /** When true, the tenant Ingress gets a `tls:` block referencing the wildcard cert. */
   ingressTlsEnabled: boolean;
 
@@ -143,6 +166,11 @@ export function _LoadOperatorConfig(): OpenClawTenantOperatorConfig
     tenantDefaultImage: _readEnvValue<string>("TENANT_DEFAULT_IMAGE", "string"),
     defaultOpenclawVersion: _readEnvValue<string>("DEFAULT_OPENCLAW_VERSION", "string", false, "2026.6.9"),
     ingressDomain: _readEnvValue<string>("INGRESS_DOMAIN", "string"),
+    ingressIp: _readEnvValue<string>("INGRESS_IP", "string", false, ""),
+    dnsManagedZone: _readEnvValue<string>("DNS_MANAGED_ZONE", "string", false, ""),
+    certManagerIssuerName: _readEnvValue<string>("CERT_MANAGER_ISSUER_NAME", "string", false, "opencrane-issuer"),
+    certManagerIssuerKind: _readEnvValue<string>("CERT_MANAGER_ISSUER_KIND", "string", false, "ClusterIssuer") === "Issuer" ? "Issuer" : "ClusterIssuer",
+    clusterTenantNamespacePrefix: _readEnvValue<string>("CLUSTER_TENANT_NAMESPACE_PREFIX", "string", false, "opencrane-"),
     ingressTlsEnabled: _readEnvValue<boolean>("INGRESS_TLS_ENABLED", "boolean", false, false),
     ingressTlsSecretName: _readEnvValue<string>("INGRESS_TLS_SECRET_NAME", "string", false, "opencrane-wildcard-tls"),
     gatewayPort: _readEnvValue<number>("GATEWAY_PORT", "number"),
